requirements.txt
requests
python-dotenv
langchain-community
langchain-text-splitters
sentence-transformers
chromadb